--- Examples/Basic Bot/templates/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Panel</title>
-	</head>
-	<body>
-		<p>Liste des channels : </p>
-		<p>{chann}</p>
-		<form action="/send_message" method="POST">
-			<p>Numéro du Channel : <input name="channel" ></p>
-			<p>Message : <input name="message"></p>
-			<input type="submit" value="Valider">
-		</form>
-	</body>
-</html>
