--- Examples/Send_Message Panel/files/js/send_message.js
@@ -1,6 +1,0 @@
-const entry_chann = document.getElementById("channel")
-const true_chann = document.getElementById("true_channel")
-function chann(id){
-	entry_chann.value = id;
-	true_chann.value = id;
-}
