--- Examples/Send_Message Panel/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Panel</title>
-	</head>
-	<body>
-		<a href="/connect">Se connecter</a>
-		<a href="/panel">Panel</a>
-	</body>
-</html>
