--- Examples/Send_Message Panel/templates/send_message.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title></title>
-		<link href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css" rel="stylesheet">
-	</head>
-	<body class="bg-blue-200">
-		<div class="flex justify-center my-20">
-			<div class="w-full max-w-xs mx-8">
-				<p class="text-center text-sm text-red-700">{channs}</p>
-				<form class="bg-gray-100 shadow-md rounded px-8 pt-6 pb-8" method="POST">
-					<div class="text-center">
-						<p class="font-bold text-blue-500 text-3xl mb-4">Send Message</p>
-						<input name="channel" id="true_channel" type="hidden">
-						<div class="mb-4">
-							<p class="text-gray-700 text-sm font-bold mb-2">Channel ID</p>
-							<input disabled id="channel">
-						</div>
-						<div class="mb-4">
-							<p class="text-gray-700 text-sm font-bold mb-2">Message</p>
-							<input name="message">
-						</div>
-						<button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">Send</button>
-					</div>
-				</form>
-			</div>
-		</div>
-	</body>
-	<script src="/files/js/send_message.js"></script>
-</html>
