+++ Piscord/API_Elements.py
@@ -1,284 +1,3 @@
-import asyncio
-import json
-import aiohttp
-from threading import Thread
-
-class Utility:
-
-	def get_self_user(self):
-		return User(asyncio.run(self.api_call("/users/@me", "GET")), self)
-
-	def get_self_guilds(self):
-		return [Guild(guild,self) for guild in asyncio.run(self.api_call("/users/@me/guilds", "GET"))]
-
-	def send_message(self,channel,**kwargs):
-		return Message(asyncio.run(self.api_call(f"/channels/{channel}/messages", "POST", json=kwargs)),self)
-
-	def get_guild(self,guild_id):
-		return Guild(asyncio.run(self.api_call(f"/guilds/{guild_id}","GET")),self)
-
-	def get_channel(self,channel_id):
-		return Channel(asyncio.run(self.api_call(f"/channels/{channel_id}","GET")),self)
-
-	def get_user(self,user_id):
-		return User(asyncio.run(self.api_call(f"/users/{user_id}")), self)
-
-	def get_invite(self, invite_code):
-		return Invite(asyncio.run(self.api_call(f"/invites/{invite_code}","GET", params={"with_counts":"true"})),self)
-
-class Events:
-	def __init__(self):
-		
-		@self.def_event("MESSAGE_CREATE","on_message")
-		class Event(Message):
-
-			def __init__(self, bot, data):
-				Message.__init__(self, data, bot)
-
-		@self.def_event("READY","on_ready")
-		class Event(User):
-
-			def __init__(self, bot, data):
-				self.version = data["v"]
-				User.__init__(self, data["user"],bot)
-
-		@self.def_event("MESSAGE_REACTION_ADD","reaction_add")
-		class Event(Member):
-
-			def __init__(self,bot,data):
-				Member.__init__(self,{**data["member"],"guild_id":data["guild_id"]},bot)
-				self.emoji = Emoji(data["emoji"])
-				self.channel_id = data["channel_id"]
-				self.message_id = data["message_id"]
-				self.__bot = bot
-
-			def get_message(self):
-				return Message(asyncio.run(self.__bot.api_call(f"/channels/{self.channel_id}/messages/{self.message_id}")),self.__bot)
-
-			def delete(self):
-				if self.id == self.__bot.get_self_user().id:
-					self.get_message().delete_self_reaction(self.emoji.name)
-				else:
-					self.get_message().delete_reaction(self.emoji.name,user_id=self.id)
-
-		@self.def_event("MESSAGE_REACTION_REMOVE","reaction_remove")
-		class Event:
-
-			def __init__(self, bot, data):
-				self.user_id = data["user_id"]
-				self.channel_id = data["channel_id"]
-				self.message_id = data["message_id"]
-				self.guild_id = data["guild_id"]
-				self.emoji = Emoji(data["emoji"])
-				self.__bot = bot
-
-			def get_message(self):
-				return Message(asyncio.run(self.__bot.api_call(f"/channels/{self.channel_id}/messages/{self.message_id}")),self.__bot)
-
-		@self.def_event("CHANNEL_CREATE","channel_create")
-		class Event(Channel):
-
-			def __init__(self, bot, data):
-				Channel.__init__(self, data, bot)
-
-		@self.def_event("CHANNEL_DELETE","channel_delete")
-		class Event(Channel):
-
-			def __init__(self, bot, data):
-				Channel.__init__(self, data, bot)
-
-		@self.def_event("CHANNEL_UPDATE","channel_update")
-		class Event(Channel):
-
-			def __init__(self, bot, data):
-				Channel.__init__(self, data, bot)
-
-		@self.def_event("GUILD_MEMBER_ADD","member_join")
-		class Event(Member):
-
-			def __init__(self, bot, data):
-				Member.__init__(self,data,bot)
-
-		@self.def_event("GUILD_MEMBER_REMOVE","member_quit")
-		class Event(User):
-
-			def __init__(self, bot, data):
-				self.guild_id = data["guild_id"]
-				User.__init__(self,data["user"],bot)
-
-		@self.def_event("GUILD_ROLE_CREATE","role_create")
-		class Event(Role):
-
-			def __init__(self, bot, data):
-				Role.__init__(self, data, bot)
-
-		@self.def_event("GUILD_ROLE_DELETE","role_delete")
-		class Event:
-
-			def __init__(self, bot, data):
-				self.guild_id = data["guild_id"]
-				self.role_id = data["role_id"]
-
-		@self.def_event("INVITE_CREATE","invite_create")
-		class Event(Invite):
-
-			def __init__(self, bot, data):
-				self.channel_id = data["channel_id"]
-				self.guild_id = data.get("guild_id",None)
-				Invite.__init__(self, data, bot)
-
-		@self.def_event("INVITE_DELETE","invite_delete")
-		class Event:
-
-			def __init__(self, bot, data):
-				self.code = data["code"]
-				self.channel_id = data["channel_id"]
-				self.guild_id = data.get("guild_id",None)
-
-class Bot(Thread,Utility,Events):
-
-	def __init__(self,token):
-		self.events_list = {}
-		Events.__init__(self)
-		Thread.__init__(self)
-		Utility.__init__(self)
-		self.token=token
-		self.api="https://discordapp.com/api"
-		self.__last_sequence = ""
-		self.events = {}
-
-	def def_event(self,event,name):
-		def add_event(function):
-			self.events_list[event] = [name,function]
-			def thing():...
-			return thing
-		return add_event
-
-	def event(self,arg):
-		def truc():...
-
-		def add_event(function):
-			self.events[arg]=function
-			return truc
-
-		if type(arg) == str:
-			return add_event
-
-		self.events[arg.__name__] = arg
-		return truc
-
-	async def api_call(self,path, method="GET", **kwargs):
-		defaults = {
-			"headers": {
-				"Authorization": f"Bot {self.token}",
-				"User-Agent": "Bot"
-			}
-		}
-		kwargs = dict(defaults, **kwargs)
-		async with aiohttp.ClientSession() as session:
-			async with session.request(method, self.api+path,**kwargs) as response:
-				try:
-					assert 200 == response.status, response.reason
-					return await response.json()
-				except:...
-
-	async def begin(self):
-		response = await self.api_call("/gateway")
-		await self.__main(response["url"])
-
-	async def __main(self,url):
-		events = self.events_list
-		async with aiohttp.ClientSession() as session:
-			async with session.ws_connect(f"{url}?v=6&encoding=json") as ws:
-				async for msg in ws:
-					data = json.loads(msg.data)
-
-					if data["op"] == 10:
-						asyncio.create_task(self.__heartbeat(ws,data["d"]["heartbeat_interval"]))
-						await ws.send_json({
-							"op": 2,
-							"d": {
-								"token": self.token,
-								"properties": {},
-								"compress": False,
-								"large_threshold": 250
-						}})
-					elif data["op"] == 0:
-						if data["t"] in events:
-							event = events[data["t"]]
-							if event[0] in self.events:
-								x = Thread(target=self.events[event[0]],args=(event[1](self,data["d"]),))
-								x.start()
-						self.__last_sequence=data["s"]
-
-	async def __heartbeat(self,ws, interval):
-		while True:
-			await asyncio.sleep(interval / 1000)
-			await ws.send_json({"op": 1,"d": self.__last_sequence})
-
-	def run(self):
-		self.loop = asyncio.new_event_loop()
-		try:
-			self.loop.run_until_complete(self.begin())
-		except RuntimeError:print("Bot éteint")
-
-	def stop(self):
-		try:
-			self.loop.stop()
-		except:...
-		try:
-			for x in asyncio.all_tasks():
-				try:
-					x.cancel()
-				except:...
-		except:...
-
-class OAuth:
-
-	def __init__(self, bot, secret, redirect_uri, scope):
-		self.bot = bot
-		self.secret = secret
-		self.id = bot.get_self_user().id
-		self.redirect_uri = redirect_uri
-		self.scope = scope
-
-	def get_url(self):
-		return f"https://discordapp.com/api/oauth2/authorize?client_id={self.id}&redirect_uri={self.redirect_uri}&response_type=code&scope={'%20'.join(self.scope.split())}"
-
-	def get_token(self, code):
-		data = {
-			"client_id": self.id,
-			"client_secret": self.secret,
-			"grant_type": "authorization_code",
-			"code": code,
-			"redirect_uri": self.redirect_uri,
-			"scope": self.scope
-		}
-
-		return asyncio.run(self.bot.api_call("/oauth2/token","POST",data=data))
-
-	def __request_token(self,token,url):
-		headers = {
-			"Authorization": f"Bearer {json.loads(token)['access_token']}"
-		}
-
-		return asyncio.run(self.bot.api_call(url,"GET",headers=headers))
-
-	def get_user(self,token):
-		if "identify" in self.scope:
-			return User(self.__request_token(token,"/users/@me"),self.bot)
-		return "Invalid Scope"
-
-	def get_guilds(self,token):
-		if "guilds" in self.scope:
-			return [Guild(guild,self.bot) for guild in self.__request_token(token,"/users/@me/guilds")]
-		return "Invalid Scope"
-
-	def add_guild_member(self, token, guild_id, user_id):
-		if "guilds.join" in self.scope:
-			return Member({**self.bot.api_call(f"/guilds/{guild_id}/members/{user_id}","PUT",json=json.loads(token)["access_token"]),"guild_id":guild_id})
-		return "Invalid Scope"
-
 class API_Element:
 
 	def to_json(self):
@@ -377,11 +96,11 @@ class Guild(API_Element):
 		return Ban(asyncio.run(self.__bot.api_call(f"/guilds/{self.id}/bans/{user_id}")),self.__bot)
 
 	def create_channel(self,**kwargs):
-		''' kwargs : https://discordapp.com/developers/docs/resources/guild#create-guild-channel '''
+		''' kwargs : https://discord.com/developers/docs/resources/guild#create-guild-channel '''
 		return Channel(asyncio.run(self.__bot.api_call(f"/guilds/{self.id}/channels", "POST", json=kwargs)),self.__bot)
 
 	def create_role(self,**kwargs):
-		''' kwargs : https://discordapp.com/developers/docs/resources/guild#create-guild-role '''
+		''' kwargs : https://discord.com/developers/docs/resources/guild#create-guild-role '''
 		return Role(asyncio.run(self.__bot.api_call(f"/guilds/{self.id}/roles", "POST", json=kwargs)),self.__bot)
 
 class Channel(API_Element):
@@ -508,7 +227,11 @@ class User(API_Element):
 		self.discriminator = user["discriminator"]
 		self.avatar = None
 		if user["avatar"]:
-			self.avatar = f"https://cdn.discordapp.com/avatars/{self.id}/{user['avatar']}.png"
+			if user["avatar"].startswith("a_"):
+				avatar_type = ".gif"
+			else:
+				avatar_type = ".png"
+			self.avatar = f"https://cdn.discordapp.com/avatars/{self.id}/{user['avatar']}{avatar_type}"
 		self.mention = f"<@{self.id}>"
 		self.__bot = bot
 
@@ -728,48 +451,3 @@ class Overwrite(API_Element):
 
 	def delete(self):
 		asyncio.run(self.__bot.api_call(f"/channels/{self.channel.id}/permissions/{self.id}","DELETE"))
-
-class Perm:
-
-	def __init__(self,number):
-		self.number = 2**number
-
-	def __eq__(self,n):
-		if n & self.number == self.number:
-			return True
-		else:
-			return False
-
-class Permission:
-
-	create_instant_invite = Perm(0)
-	kick_members = Perm(1)
-	ban_members = Perm(2)
-	administrator = Perm(3)
-	manage_channels = Perm(4)
-	manage_guild = Perm(5)
-	add_reactions = Perm(6)
-	view_audit_log = Perm(7)
-	priority_speaker = Perm(8)
-	stream = Perm(9)
-	view_channel = Perm(10)
-	send_messages = Perm(11)
-	send_tts_messages = Perm(12)
-	manage_messages = Perm(13)
-	embed_links = Perm(14)
-	attach_files = Perm(15)
-	read_message_history = Perm(16)
-	mention_everyone = Perm(17)
-	use_external_emojis = Perm(18)
-	view_guild_insights = Perm(19)
-	connect = Perm(20)
-	speak = Perm(21)
-	mute_members = Perm(22)
-	deafen_members = Perm(23)
-	move_members = Perm(24)
-	use_vad = Perm(25)
-	change_nickname = Perm(26)
-	manage_nicknames = Perm(27)
-	manage_roles = Perm(28)
-	manage_webhooks = Perm(29)
-	manage_emojis = Perm(30)
