+++ Piscord/__init__.py
@@ -1,0 +1,4 @@
+from .Piscord import *
+from .API_Elements import *
+from .OAuth import *
+from . import Permission
